code source/shopina-env/backend/requirements.txt
Django==5.2.7
djangorestframework>=3.14.0
djangorestframework-simplejwt>=2.9.0
django-cors-headers>=4.0.0
django-filter>=23.1
django-allauth>=0.61.0
pytest
pytest-django
stripe
Pillow>=10.0.0
drf-spectacular>=0.27.0
python-decouple>=3.8
